refactor(frontend): extract shared/ (authStore, sessionSchema, useSessionSync, Field)
Step 5 of #57: prep for App.tsx split.

Move 4 cross-cutting concerns out of App.tsx into shared/ so they can
be imported by future feature/ folders:
- shared/auth/authStore.ts: zustand user store + User type
- shared/auth/sessionSchema.ts: zod session envelope schema
- shared/auth/useSessionSync.ts: query + useEffect that mirrors server
  session into the client store
- shared/ui/Field.tsx: Label + control wrapper

Net effect on App.tsx: -42 lines.
