+++ frontend/src/shared/auth/authStore.ts
@@ -1,0 +1,13 @@
+import { create } from "zustand";
+
+export type User = { id: number; email: string; name: string | null };
+
+type AuthState = {
+  user: User | null;
+  setUser: (user: User | null) => void;
+};
+
+export const useAuthStore = create<AuthState>((set) => ({
+  user: null,
+  setUser: (user) => set({ user })
+}));
